requirements.txt
python-dotenv
praw
pysqlite3